+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # v2.4.24
 BUG FIXES
-- fix `check_parts_columns` corner cases for AggregateFunction versioning, fix [819](https://github.com/Altinity/clickhouse-backup/issues/819)
+- fix `check_parts_columns` corner cases for `AggregateFunction` and `SimpleAggregateFunction` versioning, fix [819](https://github.com/Altinity/clickhouse-backup/issues/819)
 
 # v2.4.23
 IMPROVEMENTS